refactor : RoomApi, RoomRestApi 분리
